Remove NOVA_RAX_AUTH hack
That Rackspace's beta Nova/Keystone does not properly work with
a stock OpenStack client is not a bug in the OpenStack client, and
should not have special case flags in the OpenStack codebase. If we
accomodate one vendor with special-case flags in the mainline codebase,
we will need to accomodate any other non-standard diversion that any
company dreams up.

Fixes bug 966329

Change-Id: Idc3ef4ae612ccd65d2ddc2891aa0f09ba1b41605